Finish d13 part1 - refactor

+++ 2024/d13/src/main/java/org/example/Button.java
@@ -4,9 +4,9 @@ public class Button {
 
     String id;
     Pair xy;
-    int counter = 0;
+    long counter = 0;
 
-    int tokenCounter, tokenPrice;
+    long tokenCounter, tokenPrice;
 
     public Button(String id, Pair xy, int tokenCounter) {
         this.id = id;
