+++ 2024/d13/src/main/java/org/example/Pair.java
@@ -1,4 +1,4 @@
 package org.example;
 
-public record Pair(int x, int y) {
+public record Pair(long x, long y) {
 }
